Disallow cancelled failed workflows
Previously, we allowed cancelling failed workflows as a workaround for
"resolving". We cancelled a failed job in order to manually acknowledge
that we were done investigating that job.

We recently added the concept ResolvedByUser in workflow-manager:
https://github.com/Clever/workflow-manager/pull/101

This means we can remove the overloaded idea of cancelling failed jobs,
and instead mark them as ResolvedByUser.