update loop should not look at jobs array when deciding to requeue
I took a peek at the SQS queue and found some workflows that were
really old, but still in the queue. Turns out these workflows were
finished, but had jobs that were still "in progress". This would
happen if someone loaded it in hubble (triggering an update of the
jobs array), but then never again loaded it.

The workflow.Jobs array is kind of a dumpster fire right now, so I'm
not entirely surprised by this bug... in the long run I think it
should be removed from the workflow object.