Relax version requirement on `github.com/golang/mock`
This is causing problems in apps that depend on both App Service and
Workflow Manager.

App Service depends on the latest stable release of
`github.com/golang/mock` which is currently `1.2.0`. Workflow Manger
depends on an even newer version, `master`.

The latest stable release (`1.2.0`) will satisfy the requirements of
App Service, Workflow Manager, and all of the apps that depend on both
Workflow Manager and App Service.

    Solving failure: No versions of github.com/Clever/workflow-manager met constraints:
      0.9.3: Could not introduce github.com/Clever/workflow-manager
      (from git@github.com:Clever/workflow-manager.git)@0.9.3,
      as it has a dependency on github.com/golang/mock with constraint
      master, which has no overlap with the following existing constraints:
      ^1.2.0 from github.com/Clever/app-service
      (from git@github.com:Clever/app-service.git)